# Server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Billu Bazaar API Documentation",
    "description": "Comprehensive OpenAPI 3.0 Swagger specification for all backend API endpoints and CRUD operations.",
    "version": "1.0.0",
    "contact": {
      "name": "Billu Bazaar API Team",
      "email": "harish05082004@gmail.com"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative Path / Current Domain (Best for HTTPS & Live)"
    },
    {
      "url": "https://demo3api.freshmindz.in",
      "description": "Environment API Server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/variants/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "origin",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/variants/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "origin",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide JWT Token in format: Bearer <token>"
      }
    },
...
  }
}